[feat][mesh] mesh 能力提升

+++ jcommon/docean-plugin/docean-plugin-dmesh/src/main/java/com/xiaomi/youpin/docean/plugin/dmesh/ms/Nacos.java
@@ -25,7 +25,8 @@ import com.xiaomi.youpin.docean.plugin.dmesh.anno.MeshMsService;
  */
 @MeshMsService(interfaceClass = Nacos.class, name = "nacos")
 public interface Nacos {
-    String getConfigStr(String dataId, String group, int timeout);
+
+    String getConfigStr(String dataId, String group, long timeout);
 
     boolean publishConfig(String dataId, String group, String content);
